reduce code lines in routes files
