+++ app/processing/debt/index.js
@@ -1,3 +1,4 @@
+const { captureDebtData } = require('./capture-debt-data')
 const { saveDebtData } = require('./save-debt-data')
 
-module.exports = { saveDebtData }
+module.exports = { captureDebtData, saveDebtData }
