+++ app/processing/enrich/index.js
@@ -1,0 +1,3 @@
+const { enrichRequest } = require('./enrich-request')
+
+module.exports = { enrichRequest }
