+++ app/routes/enrich-request.js
@@ -1,16 +1,12 @@
-const { getPaymentRequestByInvoiceNumberAndRequestId, updatePaymentRequestCategory } = require('../payment-request')
-const { saveDebt } = require('../debt')
-const { updateQualityChecksStatus } = require('../quality-check')
+const { getPaymentRequestByInvoiceNumberAndRequestId } = require('../payment-request')
 const format = require('../utils/date-formatter')
 const ViewModel = require('./models/enrich-request')
 const enrichRequestSchema = require('./schemas/enrich-request')
 const dateSchema = require('./schemas/date')
 const { enrichment } = require('../auth/permissions')
 const { getUser } = require('../auth')
-const { PENDING, PASSED } = require('../quality-check/statuses')
-const { LEDGER_CHECK, LEDGER_ENRICHMENT } = require('../payment-request/categories')
-const { sendEnrichRequestEvent } = require('../event')
-const { checkAwaitingManualLedgerDebtData } = require('../manual-ledger')
+const { enrichRequest } = require('../processing/enrich')
+
 module.exports = [{
   method: 'GET',
   path: '/enrich-request',
@@ -88,34 +84,9 @@ module.exports = [{
         return h.redirect('/enrich')
       }
 
-      if (paymentRequest.categoryId === LEDGER_ENRICHMENT) {
-        await updatePaymentRequestCategory(paymentRequestId, LEDGER_CHECK)
-        await updateQualityChecksStatus(paymentRequestId, PENDING)
-      }
-
       const user = getUser(request)
-      const { schemeId, frn } = paymentRequest
-      await saveDebt({
-        paymentRequestId: paymentRequestId,
-        schemeId: schemeId,
-        frn: frn,
-        reference: paymentRequest.agreementNumber,
-        netValue: paymentRequest.netValue ?? paymentRequest.value,
-        debtType: payload['debt-type'],
-        recoveryDate: `${day}/${month}/${year}`,
-        createdDate: new Date(),
-        createdBy: user.username,
-        createdById: user.userId
-      })
+      await enrichRequest(user, payload, paymentRequest)
 
-      const isAwaitingManualLedgerDebtData = await checkAwaitingManualLedgerDebtData(paymentRequestId)
-      if (isAwaitingManualLedgerDebtData) {
-        await updatePaymentRequestCategory(paymentRequestId, LEDGER_CHECK)
-        await updateQualityChecksStatus(paymentRequestId, PASSED)
-      } else {
-        await updateQualityChecksStatus(paymentRequestId, PENDING)
-        await sendEnrichRequestEvent(paymentRequest, user)
-      }
       return h.redirect('/enrich')
     }
   }
